Lowkey screwed up terrain mapping but it's okay
I'm going to use this branch to work on new physics, so that characters actually use the 2D space to walk around, and then jumping will be some non-physics mechanic that will allow me to cross pits in the floor and stuff. Consider "dist_from_ground", and keeping track of the y-coord prior to jumping, so that you land in the same place.